- Testing state controlled code
This iteration of code manipulates subsystems based on the state of the robot instead of the controller inputs directly.  In turn the controller now affects the state, (what the robot is currently doing). This addition adds another layer of abstraction between subsystems and user inputs and allows for complex automation  processes to be completed via simply manipulating state values.